\name{traceRkinetic-package}
\alias{traceRkinetic-package}
\alias{traceRkinetic}
\docType{package}

\title{
Kinetic analysis for dynamic nuclear imaging: compartmental models and linear
approximations.
}

\description{
This package implements compartmental models and linear 
approximations needed to perform kinetic analysis on dynamic medical images.
The provided functions use the tissue time-activity curves, the input function
and the frame times as input, and return the kinetic parameters and the 
estimated errors.
}

\details{
    \tabular{ll}{
    Package: \tab traceRkinetic\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2013-04-12\cr
    License: \tab GNU GPL 3.0\cr
    }
}

\author{
Jose Maria Mateos-Perez <\email{jmmateos@mce.hggm.es}>
}

\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[traceRkinetic:traceRkinetic-package]{<traceRkinetic>}} ~~
}

\examples{
## sExamples
}
